backend/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
python-dotenv
psycopg2-binary
psycopg2-binary

